pyproject.toml
[tool.poetry]
name = "file-processor"
version = "0.1.0"
description = "A file processor application that extracts metadata and stores it in a database."
authors = ["Mikolaj Henklewski <m.henklewski@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9.11"
pyspark = "3.5.0"
psycopg2 = "^2.9.7"
pefile = "^2023.2.7"
boto3 = "^1.28.43"
yoyo-migrations = "^8.2.0"
pydantic = "^2.4.2"
asyncpg = "^0.28.0"
pydantic-settings = "^2.0.3"
black = "^23.11.0"
flake8 = "^6.1.0"
mypy = "^1.7.1"
isort = "^5.12.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.23.2"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
line_length = 88
multi_line_output = 3
lines_after_imports = 2
include_trailing_comma = true
balanced_wrapping = true
profile = "black"

[tool.mypy]
plugins = "pydantic.mypy"
ignore_missing_imports = true
disallow_untyped_defs = true

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]

[tool.black]
line-length = 88